[EV-659] Moved dashcore RPC client to separate file
