+++ lib/api/dashcore/ZmqClient.js
