+++ lib/app.js
@@ -3,7 +3,7 @@ const config = require('./config');
 const log = require('./log');
 const rpcServer = require('./rpcServer');
 const quorumService = require('./services/quorum');
-const ZmqClient = require('./api/dashcore/zmq');
+const ZmqClient = require('./api/dashcore/ZmqClient');
 
 async function main() {
   /* Application start */
