+++ lib/services/quorum/index.js
@@ -1,5 +1,5 @@
 const qDash = require('quorums-dash');
-const dashcore = require('../../api/dashcore');
+const dashcore = require('../../api/dashcore/rpc');
 const logger = require('../../log');
 const insight = require('../../api/insight');
